+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/abstracts/Request.kt
@@ -1,6 +1,6 @@
 package com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts
 
-import com.github.insanusmokrassar.TelegramBotAPI.types.ResponseParameters
+import com.github.insanusmokrassar.TelegramBotAPI.types.Response
 import com.github.insanusmokrassar.TelegramBotAPI.utils.toJsonWithoutNulls
 import kotlinx.serialization.*
 import kotlinx.serialization.json.JsonObject
@@ -15,6 +15,6 @@ interface Request<T: Any> {
 fun <T : Any> StringFormat.extractResult(
     from: String,
     dataSerializer: KSerializer<T>
-): ResponseParameters<T> {
-    return parse(ResponseParameters.serializer(dataSerializer), from)
+): Response<T> {
+    return parse(Response.serializer(dataSerializer), from)
 }